Securing the Views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
 from django.db.models import Q
 # Q returns 'or' logic for queries
@@ -95,10 +96,16 @@ def product_detail(request, product_id):
     # context returns things back to template
 
 
+@login_required
 def add_product(request):
     """
     Add a product to the store
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         # instantiate a new instance of the form
         # to capture the image, if one is added
@@ -122,10 +129,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit a product in the store
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse('home'))
 
     # get the product
     product = get_object_or_404(Product, pk=product_id)
@@ -154,10 +166,16 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete a product from the store
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse('home'))
+    
     # get product (or 404)
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
